3.4-django-testing/django_testing/requirements.txt
django~=3.2.9
djangorestframework~=3.12.4
psycopg2-binary

pytest~=6.2.5
django-filter~=21.1